backend for chatting: change word from room to session
